fix: correct MiniPay stack detection and trim the Mini App bundle
useIsMiniPay derived its answer from window.ethereum in an effect, so it was
false on the first render: inside MiniPay the create-Stack modal mounted the
Privy variant for one commit, calling Privy hooks outside PrivyProvider and
starting EIP-712 invite generation — the one flow MiniPay cannot do.

- Publish isMiniPay from the provider branch itself (new is-minipay context)
  so consumers are correct on render #1; drop the effect-based hook.
- Only fall back to the Privy stack after ethereum#initialized or a short
  grace period, so a late-injected provider no longer strands a real MiniPay
  session on the wrong stack.
- Wait for receipts on the mounted wagmi config (useConfig) instead of
  importing the Privy/RainbowKit one directly.
- Derive the removable-member list from getCircle().owner rather than
  assuming the owner is still at index 0.
- Code-split both provider stacks (extract PrivyProviders, next/dynamic
  both): the layout chunk no longer carries Privy, RainbowKit or the Celo
  wagmi config, so each client downloads only the stack it uses.
- Bundle the RainbowKit stylesheet instead of fetching it from unpkg on
  every page — a render-blocking third-party request and an extra origin.
- Convert the hero image to WebP (117KB -> 58KB) and Celo-gate the onboarding
  copy that still told Celo users to send BREAD or xDAI.
- Document the MiniPay build, network manifest and submission checklist in
  docs/MINIPAY.md, including why exact dependency pinning and
  ignore-scripts=true are deliberately not adopted.

Refs #154
