+++ src/app/account/_components/account-overview-card.tsx
@@ -14,7 +14,7 @@ import { Address } from "viem";
 import LocalButton from "@/components/button";
 import { useIsOwnAddress } from "@/hooks/use-is-own-address";
 import { formatDepositAmount } from "@/lib/deposit-token";
-import { useIsMiniPay } from "@/hooks/use-is-minipay";
+import { useIsMiniPay } from "@/components/providers/is-minipay";
 import { MINIPAY_ADD_CASH_URL } from "@/utils/minipay";
 
 const toBread = (value: bigint | undefined) =>
