+++ src/hooks/use-wait-for-tx-receipt.ts
@@ -1,11 +1,15 @@
 import { Address } from "viem";
+import { useConfig } from "wagmi";
 import { waitForTransactionReceipt } from "@wagmi/core";
-import { wagmiConfig } from "@/components/providers/web3";
 import { getDefaultChainId } from "@/utils/chain";
 
 export const useWaitForTxReceipt = () => {
+  // Whichever config the mounted provider stack supplies — the Privy one on
+  // the web build, the injected-wallet one inside MiniPay.
+  const config = useConfig();
+
   const waitForTxReceipt = async (hash: Address) => {
-    await waitForTransactionReceipt(wagmiConfig, {
+    await waitForTransactionReceipt(config, {
       hash,
       chainId: getDefaultChainId(),
     });
